+++ app/Scraper/Manager.php
@@ -21,6 +21,7 @@ final class Manager
         \Kami\Cocktail\Scraper\Sites\LiberAndCo::class,
         \Kami\Cocktail\Scraper\Sites\CocktailsDistilled::class,
         \Kami\Cocktail\Scraper\Sites\PunchDrink::class,
+        \Kami\Cocktail\Scraper\Sites\SteveTheBartender::class,
     ];
 
     public function __construct(private readonly string $url)