+++ app/UnitConverter/Dash.php
@@ -6,8 +6,8 @@ namespace Kami\Cocktail\UnitConverter;
 
 class Dash extends Unit
 {
-    public function toMl(): Ml
-    {
-        return new Ml($this->getValue() * 0.3125);
-    }
+    // public function toMl(): Ml
+    // {
+    //     return new Ml($this->getValue() * 0.3125);
+    // }
 }